Complete re-write
This re-write simplifies the API dramatically.
Notably, the concept of pipes is gone. I am no longer convinced it is
worth it: A pipe "failure" is a bit more nuanced than i originally
thought. There are at least two types of "failure":
(1) A pipe is not even interested in a request
(2) A pipe is interested in a request, but failed to process it.

No distinction was made between these in the previous implementation.
Instead, I leaned on users of the library using the right combination of
uhh...combinators to capture/ignore the right failures. This means a lot
of symbols imported, which makes the crate harder to use.

I realized that there is no need for generality in the form of pipe
combinators. There is basically one correct way all the pipes fit
together...so why not just hard-code it....and that's what i did